requirements.txt
# requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
python-jose